Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -5,8 +5,8 @@ char* welcome() {
 }
 
 int main() {
-    float a;
-    scanf("%f",&a);
-    printf("you entered:%f ",a);
+    int a;
+    scanf("%d",&a);
+    printf("you entered:%d ",a);
     return 0;
 }
